extract IntentDecoder to reuse
